+++ src/platform/plugins/shared/discover/public/application/main/components/session_view/discover_main_app.tsx
@@ -9,17 +9,17 @@
 
 import React, { useEffect } from 'react';
 import { RootDragDropProvider } from '@kbn/dom-drag-drop';
-import { useUrlTracking } from './hooks/use_url_tracking';
-import type { DiscoverStateContainer } from './state_management/discover_state';
-import { DiscoverLayout } from './components/layout';
-import { setBreadcrumbs } from '../../utils/breadcrumbs';
-import { addHelpMenuToAppChrome } from '../../components/help_menu/help_menu_util';
-import { useDiscoverServices } from '../../hooks/use_discover_services';
-import { useSavedSearchAliasMatchRedirect } from '../../hooks/saved_search_alias_match_redirect';
-import { useSavedSearchInitial } from './state_management/discover_state_provider';
-import { useAdHocDataViews } from './hooks/use_adhoc_data_views';
-import { useEsqlMode } from './hooks/use_esql_mode';
-import { addLog } from '../../utils/add_log';
+import { useUrlTracking } from '../../hooks/use_url_tracking';
+import type { DiscoverStateContainer } from '../../state_management/discover_state';
+import { DiscoverLayout } from '../layout';
+import { setBreadcrumbs } from '../../../../utils/breadcrumbs';
+import { addHelpMenuToAppChrome } from '../../../../components/help_menu/help_menu_util';
+import { useDiscoverServices } from '../../../../hooks/use_discover_services';
+import { useSavedSearchAliasMatchRedirect } from '../../../../hooks/saved_search_alias_match_redirect';
+import { useSavedSearchInitial } from '../../state_management/discover_state_provider';
+import { useAdHocDataViews } from '../../hooks/use_adhoc_data_views';
+import { useEsqlMode } from '../../hooks/use_esql_mode';
+import { addLog } from '../../../../utils/add_log';
 
 const DiscoverLayoutMemoized = React.memo(DiscoverLayout);
