+++ src/platform/plugins/shared/discover/public/application/main/components/session_view/session_view.tsx
@@ -1,0 +1,382 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import React, { useEffect, useState } from 'react';
+import type { IKbnUrlStateStorage } from '@kbn/kibana-utils-plugin/public';
+import { SavedObjectNotFound } from '@kbn/kibana-utils-plugin/public';
+import { useParams } from 'react-router-dom';
+import useLatest from 'react-use/lib/useLatest';
+import { isOfAggregateQueryType } from '@kbn/es-query';
+import { getSavedSearchFullPathUrl } from '@kbn/saved-search-plugin/public';
+import { i18n } from '@kbn/i18n';
+import type { DataView } from '@kbn/data-views-plugin/common';
+import { cloneDeep, isEqual } from 'lodash';
+import useAsyncFn from 'react-use/lib/useAsyncFn';
+import { useExecutionContext } from '@kbn/kibana-react-plugin/public';
+import { getValidFilters } from '../../../../utils/get_valid_filters';
+import { isRefreshIntervalValid, isTimeRangeValid } from '../../../../utils/validate_time';
+import { useUrl } from '../../hooks/use_url';
+import { useAlertResultsToast } from '../../hooks/use_alert_results_toast';
+import { createDataViewDataSource, isDataViewSource } from '../../../../../common/data_sources';
+import { copySavedSearch } from '../../state_management/discover_saved_search_container';
+import { loadAndResolveDataView } from '../../state_management/utils/resolve_data_view';
+import { setBreadcrumbs } from '../../../../utils/breadcrumbs';
+import type { MainHistoryLocationState } from '../../../../../common';
+import { useDiscoverServices } from '../../../../hooks/use_discover_services';
+import { getEsqlDataView } from '../../state_management/utils/get_esql_data_view';
+import {
+  APP_STATE_URL_KEY,
+  getInitialState,
+} from '../../state_management/discover_app_state_container';
+import type {
+  AppStateUrl,
+  DiscoverAppState,
+} from '../../state_management/discover_app_state_container';
+import { getDiscoverStateContainer } from '../../state_management/discover_state';
+import type { DiscoverStateContainer } from '../../state_management/discover_state';
+import {
+  RuntimeStateProvider,
+  internalStateActions,
+  useInternalStateDispatch,
+  useInternalStateSelector,
+  useRuntimeState,
+} from '../../state_management/redux';
+import type {
+  CustomizationCallback,
+  DiscoverCustomizationContext,
+} from '../../../../customizations';
+import type { InternalStateStore, RuntimeStateManager } from '../../state_management/redux';
+import type { RootProfileState } from '../../../../context_awareness';
+import {
+  DiscoverCustomizationProvider,
+  useDiscoverCustomizationService,
+} from '../../../../customizations';
+import { cleanupUrlState } from '../../state_management/utils/cleanup_url_state';
+import { DiscoverError } from '../../../../components/common/error_alert';
+import { NoDataPage } from './no_data_page';
+import { DiscoverMainProvider } from '../../state_management/discover_state_provider';
+import { updateSavedSearch } from '../../state_management/utils/update_saved_search';
+import { BrandedLoadingIndicator } from './branded_loading_indicator';
+import { RedirectWhenSavedObjectNotFound } from './redirect_not_found';
+import { DiscoverMainApp } from './discover_main_app';
+import type { MainInitializationState, NarrowAsyncState } from '../../types';
+
+interface DiscoverSessionViewProps {
+  rootProfileState: Extract<RootProfileState, { rootProfileLoading: false }>;
+  mainInitializationState: MainInitializationState;
+  customizationContext: DiscoverCustomizationContext;
+  customizationCallbacks: CustomizationCallback[];
+  urlStateStorage: IKbnUrlStateStorage;
+  internalState: InternalStateStore;
+  runtimeStateManager: RuntimeStateManager;
+}
+
+type SessionInitializationState =
+  | {
+      showNoDataPage: true;
+      stateContainer: undefined;
+    }
+  | {
+      showNoDataPage: false;
+      stateContainer: DiscoverStateContainer;
+    };
+
+type InitializeSession = (
+  defaultUrlState?: DiscoverAppState
+) => Promise<SessionInitializationState>;
+
+export const DiscoverSessionView = ({
+  rootProfileState,
+  mainInitializationState,
+  customizationContext,
+  customizationCallbacks,
+  urlStateStorage,
+  internalState,
+  runtimeStateManager,
+}: DiscoverSessionViewProps) => {
+  const dispatch = useInternalStateDispatch();
+  const services = useDiscoverServices();
+  const {
+    core,
+    chrome,
+    ebtManager,
+    savedSearch,
+    toastNotifications,
+    uiSettings,
+    history,
+    getScopedHistory,
+  } = services;
+  const { id: discoverSessionId } = useParams<{ id: string }>();
+  const [historyLocationState] = useState(
+    () => getScopedHistory<MainHistoryLocationState>()?.location.state
+  );
+  const defaultProfileAdHocDataViewIds = useInternalStateSelector(
+    (state) => state.defaultProfileAdHocDataViewIds
+  );
+  const initialize = useLatest<InitializeSession>(async (defaultUrlState) => {
+    internalState.dispatch(internalStateActions.resetOnSavedSearchChange());
+
+    const discoverSessionLoadTracker = ebtManager.trackPerformanceEvent('discoverLoadSavedSearch');
+    const urlState = cleanupUrlState(
+      urlStateStorage.get<AppStateUrl>(APP_STATE_URL_KEY) ?? defaultUrlState,
+      uiSettings
+    );
+    const persistedDiscoverSession = discoverSessionId
+      ? await savedSearch.get(discoverSessionId)
+      : undefined;
+    const initialQuery =
+      urlState?.query ?? persistedDiscoverSession?.searchSource.getField('query');
+    const isEsqlMode = isOfAggregateQueryType(initialQuery);
+    const discoverSessionDataView = persistedDiscoverSession?.searchSource.getField('index');
+    const discoverSessionHasAdHocDataView = Boolean(
+      discoverSessionDataView && !discoverSessionDataView.isPersisted()
+    );
+    const profileDataViews = runtimeStateManager.adHocDataViews$
+      .getValue()
+      .filter(({ id }) => id && defaultProfileAdHocDataViewIds.includes(id));
+    const profileDataViewsExist = profileDataViews.length > 0;
+    const locationStateHasDataViewSpec = Boolean(historyLocationState?.dataViewSpec);
+    const canAccessWithoutPersistedDataView =
+      isEsqlMode ||
+      discoverSessionHasAdHocDataView ||
+      profileDataViewsExist ||
+      locationStateHasDataViewSpec;
+
+    if (!mainInitializationState.hasUserDataView && !canAccessWithoutPersistedDataView) {
+      return { showNoDataPage: true };
+    }
+
+    if (customizationContext.displayMode === 'standalone' && persistedDiscoverSession) {
+      if (persistedDiscoverSession.id) {
+        chrome.recentlyAccessed.add(
+          getSavedSearchFullPathUrl(persistedDiscoverSession.id),
+          persistedDiscoverSession.title ??
+            i18n.translate('discover.defaultDiscoverSessionTitle', {
+              defaultMessage: 'Untitled Discover session',
+            }),
+          persistedDiscoverSession.id
+        );
+      }
+
+      setBreadcrumbs({ services, titleBreadcrumbText: persistedDiscoverSession.title });
+    }
+
+    let dataView: DataView;
+
+    if (isOfAggregateQueryType(initialQuery)) {
+      dataView = await getEsqlDataView(
+        initialQuery,
+        runtimeStateManager.currentDataView$.getValue(),
+        services
+      );
+    } else {
+      const result = await loadAndResolveDataView({
+        dataViewId: isDataViewSource(urlState?.dataSource)
+          ? urlState?.dataSource.dataViewId
+          : discoverSessionDataView?.id,
+        dataViewSpec: historyLocationState?.dataViewSpec,
+        savedSearch: persistedDiscoverSession,
+        isEsqlMode,
+        services,
+        internalState,
+        runtimeStateManager,
+      });
+
+      dataView = result.dataView;
+    }
+
+    internalState.dispatch(internalStateActions.setDataView(dataView));
+
+    if (!dataView.isPersisted()) {
+      internalState.dispatch(internalStateActions.appendAdHocDataViews(dataView));
+    }
+
+    const stateContainer = getDiscoverStateContainer({
+      services,
+      customizationContext,
+      stateStorageContainer: urlStateStorage,
+      internalState,
+      runtimeStateManager,
+    });
+    const initialState = getInitialState({
+      initialUrlState: urlState,
+      savedSearch: persistedDiscoverSession,
+      overrideDataView: dataView,
+      services,
+    });
+    const discoverSession = updateSavedSearch({
+      savedSearch: persistedDiscoverSession
+        ? copySavedSearch(persistedDiscoverSession)
+        : savedSearch.getNew(),
+      dataView,
+      state: initialState,
+      globalStateContainer: stateContainer.globalState,
+      services,
+    });
+
+    if (discoverSession.timeRestore && dataView.isTimeBased()) {
+      const { timeRange, refreshInterval } = discoverSession;
+
+      if (timeRange && isTimeRangeValid(timeRange)) {
+        services.timefilter.setTime(timeRange);
+      }
+
+      if (refreshInterval && isRefreshIntervalValid(refreshInterval)) {
+        services.timefilter.setRefreshInterval(refreshInterval);
+      }
+    }
+
+    // Cleaning up the previous state
+    services.filterManager.setAppFilters([]);
+    services.data.query.queryString.clearQuery();
+
+    // Sync global filters (coming from URL) to filter manager.
+    // It needs to be done manually here as `syncGlobalQueryStateWithUrl` is being called after this `loadSavedSearch` function.
+    const globalFilters = stateContainer.globalState?.get()?.filters;
+    const shouldUpdateWithGlobalFilters =
+      globalFilters?.length && !services.filterManager.getGlobalFilters()?.length;
+    if (shouldUpdateWithGlobalFilters) {
+      services.filterManager.setGlobalFilters(globalFilters);
+    }
+
+    // set data service filters
+    if (initialState.filters?.length) {
+      // Saved search SO persists all filters as app filters
+      services.data.query.filterManager.setAppFilters(cloneDeep(initialState.filters));
+    }
+
+    // some filters may not be valid for this context, so update
+    // the filter manager with a modified list of valid filters
+    const currentFilters = services.filterManager.getFilters();
+    const validFilters = getValidFilters(dataView, currentFilters);
+    if (!isEqual(currentFilters, validFilters)) {
+      services.filterManager.setFilters(validFilters);
+    }
+
+    // set data service query
+    if (initialState.query) {
+      services.data.query.queryString.setQuery(initialState.query);
+    }
+
+    if (!urlState && shouldUpdateWithGlobalFilters) {
+      discoverSession.searchSource.setField(
+        'filter',
+        cloneDeep(services.filterManager.getFilters())
+      );
+    }
+
+    if (persistedDiscoverSession) {
+      stateContainer.savedSearchState.set(persistedDiscoverSession);
+      stateContainer.savedSearchState.assignNextSavedSearch(discoverSession);
+    } else {
+      stateContainer.savedSearchState.set(discoverSession);
+    }
+
+    stateContainer.appState.set(initialState);
+    discoverSessionLoadTracker.reportEvent();
+
+    return { showNoDataPage: false, stateContainer };
+  });
+  const [initializationState, initializeSession] = useAsyncFn<InitializeSession>(
+    (...params) => initialize.current(...params),
+    [initialize],
+    {
+      loading: true,
+    }
+  );
+  const initializeSessionWithDefaultLocationState = useLatest(() => {
+    const scopedHistory = getScopedHistory<{ defaultState?: DiscoverAppState }>();
+    initializeSession(scopedHistory?.location.state?.defaultState);
+  });
+  const initializeSessionState = initializationState as NarrowAsyncState<
+    typeof initializationState
+  >;
+  const customizationService = useDiscoverCustomizationService({
+    customizationCallbacks,
+    stateContainer: initializeSessionState.value?.stateContainer,
+  });
+  const currentDataView = useRuntimeState(runtimeStateManager.currentDataView$);
+  const adHocDataViews = useRuntimeState(runtimeStateManager.adHocDataViews$);
+
+  useEffect(() => {
+    initializeSessionWithDefaultLocationState.current();
+  }, [discoverSessionId, initializeSessionWithDefaultLocationState]);
+
+  useUrl({
+    history,
+    savedSearchId: discoverSessionId,
+    onNewUrl: () => {
+      initializeSessionWithDefaultLocationState.current();
+    },
+  });
+
+  useAlertResultsToast({
+    isAlertResults: historyLocationState?.isAlertResults,
+    toastNotifications,
+  });
+
+  useExecutionContext(core.executionContext, {
+    type: 'application',
+    page: 'app',
+    id: discoverSessionId || 'new',
+  });
+
+  if (initializeSessionState.loading) {
+    return <BrandedLoadingIndicator />;
+  }
+
+  if (initializeSessionState.error) {
+    if (initializeSessionState.error instanceof SavedObjectNotFound) {
+      return (
+        <RedirectWhenSavedObjectNotFound
+          error={initializeSessionState.error}
+          discoverSessionId={discoverSessionId}
+        />
+      );
+    }
+
+    return <DiscoverError error={initializeSessionState.error} />;
+  }
+
+  if (initializeSessionState.value.showNoDataPage) {
+    return (
+      <NoDataPage
+        {...mainInitializationState}
+        onDataViewCreated={async (dataViewUnknown) => {
+          await dispatch(internalStateActions.loadDataViewList());
+          const dataView = dataViewUnknown as DataView;
+          initializeSession({
+            dataSource: dataView.id
+              ? createDataViewDataSource({ dataViewId: dataView.id })
+              : undefined,
+          });
+        }}
+        onESQLNavigationComplete={() => {
+          initializeSession();
+        }}
+      />
+    );
+  }
+
+  if (!customizationService || !currentDataView) {
+    return <BrandedLoadingIndicator />;
+  }
+
+  return (
+    <DiscoverCustomizationProvider value={customizationService}>
+      <DiscoverMainProvider value={initializeSessionState.value.stateContainer}>
+        <RuntimeStateProvider currentDataView={currentDataView} adHocDataViews={adHocDataViews}>
+          <rootProfileState.AppWrapper>
+            <DiscoverMainApp stateContainer={initializeSessionState.value.stateContainer} />
+          </rootProfileState.AppWrapper>
+        </RuntimeStateProvider>
+      </DiscoverMainProvider>
+    </DiscoverCustomizationProvider>
+  );
+};
