+++ src/platform/plugins/shared/discover/public/application/main/types.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import type { AsyncState } from 'react-use/lib/useAsyncFn';
+import type { RootProfileState } from '../../context_awareness';
+
+export interface MainInitializationState {
+  hasESData: boolean;
+  hasUserDataView: boolean;
+}
+
+export type InitializeMain = (
+  rootProfileState: Extract<RootProfileState, { rootProfileLoading: false }>
+) => Promise<MainInitializationState>;
+
+export type NarrowAsyncState<TState extends AsyncState<unknown>> = Exclude<
+  TState,
+  { error?: undefined; value?: undefined }
+>;
